Move some components to grouped folder
