+++ editor.planx.uk/src/@planx/components/Checklist/Public/components/grouped/GroupedChecklist.tsx
@@ -5,6 +5,7 @@ import {
   getLayout,
   groupedChecklistValidationSchema,
 } from "@planx/components/Checklist/model";
+import { PublicChecklistProps } from "@planx/components/Checklist/types";
 import Card from "@planx/components/shared/Preview/Card";
 import { CardHeader } from "@planx/components/shared/Preview/CardHeader/CardHeader";
 import { getIn, useFormik } from "formik";
@@ -13,7 +14,6 @@ import FullWidthWrapper from "ui/public/FullWidthWrapper";
 import ErrorWrapper from "ui/shared/ErrorWrapper";
 import { object } from "yup";
 
-import { PublicChecklistProps } from "../../types";
 import { GroupedChecklistOptions } from "./GroupedChecklistOptions";
 
 export const GroupedChecklist: React.FC<PublicChecklistProps> = (props) => {
